+++ file_picker/example/lib/src/file_picker_demo.dart
@@ -5,19 +5,18 @@ import 'package:file_picker/file_picker.dart';
 
 class FilePickerDemo extends StatefulWidget {
   @override
-  _FilePickerDemoState createState() => new _FilePickerDemoState();
+  _FilePickerDemoState createState() => _FilePickerDemoState();
 }
 
 class _FilePickerDemoState extends State<FilePickerDemo> {
   final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
   String _fileName;
-  String _path;
-  Map<String, String> _paths;
+  List<PlatformFile> _paths;
   String _extension;
   bool _loadingPath = false;
   bool _multiPick = false;
   FileType _pickingType = FileType.any;
-  TextEditingController _controller = new TextEditingController();
+  TextEditingController _controller = TextEditingController();
 
   @override
   void initState() {
@@ -28,97 +27,84 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
   void _openFileExplorer() async {
     setState(() => _loadingPath = true);
     try {
-      if (_multiPick) {
-        _path = null;
-        _paths = await FilePicker.getMultiFilePath(
-          type: _pickingType,
-          allowedExtensions: (_extension?.isNotEmpty ?? false)
-              ? _extension?.replaceAll(' ', '')?.split(',')
-              : null,
-        );
-      } else {
-        _paths = null;
-        _path = await FilePicker.getFilePath(
-          type: _pickingType,
-          allowedExtensions: (_extension?.isNotEmpty ?? false)
-              ? _extension?.replaceAll(' ', '')?.split(',')
-              : null,
-        );
-      }
+      _paths = (await FilePicker.instance.pickFiles(
+        type: _pickingType,
+        allowMultiple: _multiPick,
+        allowedExtensions: (_extension?.isNotEmpty ?? false) ? _extension?.replaceAll(' ', '')?.split(',') : null,
+      ))
+          ?.files;
     } on PlatformException catch (e) {
       print("Unsupported operation" + e.toString());
+    } catch (ex) {
+      print(ex);
     }
     if (!mounted) return;
     setState(() {
       _loadingPath = false;
-      _fileName = _path != null
-          ? _path.split('/').last
-          : _paths != null ? _paths.keys.toString() : '...';
+      _fileName = _paths != null ? _paths.map((e) => e.name).toString() : '...';
     });
   }
 
   void _clearCachedFiles() {
-    FilePicker.clearTemporaryFiles().then((result) {
+    FilePicker.instance.clearTemporaryFiles().then((result) {
       _scaffoldKey.currentState.showSnackBar(
         SnackBar(
           backgroundColor: result ? Colors.green : Colors.red,
-          content: Text((result
-              ? 'Temporary files removed with success.'
-              : 'Failed to clean temporary files')),
+          content: Text((result ? 'Temporary files removed with success.' : 'Failed to clean temporary files')),
         ),
       );
     });
   }
 
   void _selectFolder() {
-    FilePicker.getDirectoryPath().then((value) {
-      setState(() => _path = value);
+    FilePicker.instance.getDirectoryPath().then((value) {
+      setState(() => _paths = [value]);
     });
   }
 
   @override
   Widget build(BuildContext context) {
-    return new MaterialApp(
-      home: new Scaffold(
+    return MaterialApp(
+      home: Scaffold(
         key: _scaffoldKey,
-        appBar: new AppBar(
+        appBar: AppBar(
           title: const Text('File Picker example app'),
         ),
-        body: new Center(
-            child: new Padding(
+        body: Center(
+            child: Padding(
           padding: const EdgeInsets.only(left: 10.0, right: 10.0),
-          child: new SingleChildScrollView(
-            child: new Column(
+          child: SingleChildScrollView(
+            child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
-                new Padding(
+                Padding(
                   padding: const EdgeInsets.only(top: 20.0),
-                  child: new DropdownButton(
-                      hint: new Text('LOAD PATH FROM'),
+                  child: DropdownButton(
+                      hint: Text('LOAD PATH FROM'),
                       value: _pickingType,
                       items: <DropdownMenuItem>[
-                        new DropdownMenuItem(
-                          child: new Text('FROM AUDIO'),
+                        DropdownMenuItem(
+                          child: Text('FROM AUDIO'),
                           value: FileType.audio,
                         ),
-                        new DropdownMenuItem(
-                          child: new Text('FROM IMAGE'),
+                        DropdownMenuItem(
+                          child: Text('FROM IMAGE'),
                           value: FileType.image,
                         ),
-                        new DropdownMenuItem(
-                          child: new Text('FROM VIDEO'),
+                        DropdownMenuItem(
+                          child: Text('FROM VIDEO'),
                           value: FileType.video,
                         ),
-                        new DropdownMenuItem(
-                          child: new Text('FROM MEDIA'),
+                        DropdownMenuItem(
+                          child: Text('FROM MEDIA'),
                           value: FileType.media,
                         ),
-                        new DropdownMenuItem(
-                          child: new Text('FROM ANY'),
+                        DropdownMenuItem(
+                          child: Text('FROM ANY'),
                           value: FileType.any,
                         ),
-                        new DropdownMenuItem(
-                          child: new Text('CUSTOM FORMAT'),
+                        DropdownMenuItem(
+                          child: Text('CUSTOM FORMAT'),
                           value: FileType.custom,
                         ),
                       ],
@@ -129,87 +115,76 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
                             }
                           })),
                 ),
-                new ConstrainedBox(
-                  constraints: BoxConstraints.tightFor(width: 100.0),
+                ConstrainedBox(
+                  constraints: const BoxConstraints.tightFor(width: 100.0),
                   child: _pickingType == FileType.custom
-                      ? new TextFormField(
+                      ? TextFormField(
                           maxLength: 15,
                           autovalidate: true,
                           controller: _controller,
-                          decoration:
-                              InputDecoration(labelText: 'File extension'),
+                          decoration: InputDecoration(labelText: 'File extension'),
                           keyboardType: TextInputType.text,
                           textCapitalization: TextCapitalization.none,
                         )
-                      : new Container(),
+                      : const SizedBox(),
                 ),
-                new ConstrainedBox(
-                  constraints: BoxConstraints.tightFor(width: 200.0),
-                  child: new SwitchListTile.adaptive(
-                    title: new Text('Pick multiple files',
-                        textAlign: TextAlign.right),
-                    onChanged: (bool value) =>
-                        setState(() => _multiPick = value),
+                ConstrainedBox(
+                  constraints: const BoxConstraints.tightFor(width: 200.0),
+                  child: SwitchListTile.adaptive(
+                    title: Text('Pick multiple files', textAlign: TextAlign.right),
+                    onChanged: (bool value) => setState(() => _multiPick = value),
                     value: _multiPick,
                   ),
                 ),
-                new Padding(
+                Padding(
                   padding: const EdgeInsets.only(top: 50.0, bottom: 20.0),
                   child: Column(
                     children: <Widget>[
-                      new RaisedButton(
+                      RaisedButton(
                         onPressed: () => _openFileExplorer(),
-                        child: new Text("Open file picker"),
+                        child: Text("Open file picker"),
                       ),
-                      new RaisedButton(
+                      RaisedButton(
                         onPressed: () => _selectFolder(),
-                        child: new Text("Pick folder"),
+                        child: Text("Pick folder"),
                       ),
-                      new RaisedButton(
+                      RaisedButton(
                         onPressed: () => _clearCachedFiles(),
-                        child: new Text("Clear temporary files"),
+                        child: Text("Clear temporary files"),
                       ),
                     ],
                   ),
                 ),
-                new Builder(
+                Builder(
                   builder: (BuildContext context) => _loadingPath
                       ? Padding(
                           padding: const EdgeInsets.only(bottom: 10.0),
-                          child: const CircularProgressIndicator())
-                      : _path != null || _paths != null
-                          ? new Container(
+                          child: const CircularProgressIndicator(),
+                        )
+                      : _paths != null
+                          ? Container(
                               padding: const EdgeInsets.only(bottom: 30.0),
                               height: MediaQuery.of(context).size.height * 0.50,
-                              child: new Scrollbar(
-                                  child: new ListView.separated(
-                                itemCount: _paths != null && _paths.isNotEmpty
-                                    ? _paths.length
-                                    : 1,
+                              child: Scrollbar(
+                                  child: ListView.separated(
+                                itemCount: _paths != null && _paths.isNotEmpty ? _paths.length : 1,
                                 itemBuilder: (BuildContext context, int index) {
-                                  final bool isMultiPath =
-                                      _paths != null && _paths.isNotEmpty;
-                                  final String name = 'File $index: ' +
-                                      (isMultiPath
-                                          ? _paths.keys.toList()[index]
-                                          : _fileName ?? '...');
-                                  final path = isMultiPath
-                                      ? _paths.values.toList()[index].toString()
-                                      : _path;
+                                  final bool isMultiPath = _paths != null && _paths.isNotEmpty;
+                                  final String name =
+                                      'File $index: ' + (isMultiPath ? _paths.map((e) => e.name).toList()[index] : _fileName ?? '...');
+                                  final path = _paths.map((e) => e.path).toList()[index].toString();
 
-                                  return new ListTile(
-                                    title: new Text(
+                                  return ListTile(
+                                    title: Text(
                                       name,
                                     ),
-                                    subtitle: new Text(path),
+                                    subtitle: Text(path),
                                   );
                                 },
-                                separatorBuilder:
-                                    (BuildContext context, int index) =>
-                                        new Divider(),
+                                separatorBuilder: (BuildContext context, int index) => const Divider(),
                               )),
                             )
-                          : new Container(),
+                          : const SizedBox(),
                 ),
               ],
             ),